line 39 of workshop feedback sheet
The Hide the password field section says to find the div containing the password and give it an identifier. This is an odd way to phrase this because many devs don’t refer to the HTML ID property by its “full” name. Also, it’s a little confusing to locate the div in the file quickly with those instructions since it only has a class to begin with.
Recommend: Provide the password field div ID in the sample code out of the box and don’t make the reader add the ID, OR change the phrasing to say “add an ID property with a value of…”.